analyzer/requirements.txt
cassandra-driver==3.17.0
